requirements.txt
matplotlib
numpy
pillow
scikit-image
scipy
tensorflow==1.14.0
pyyaml
